committing refactored class, unit tests in progress...
